requirements.txt
google-api-core==1.25.0
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-cloud-speech==2.0.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
librosa==0.8.0
llvmlite==0.34.0
numpy==1.18.5
SoundFile==0.10.3.post1